+++ src/app/google/callback/page.tsx
@@ -1,1 +1,5 @@
-export { default } from '@/app/oauth/[provider]/callback/page'
+import { OAuthCallbackPage } from '@/app/oauth/[provider]/callback/page'
+
+export default function GoogleCallbackPage() {
+  return <OAuthCallbackPage providerOverride="google" />
+}