+++ src/app/kakao/callback/page.tsx
@@ -1,1 +1,5 @@
-export { default } from '@/app/oauth/[provider]/callback/page'
+import { OAuthCallbackPage } from '@/app/oauth/[provider]/callback/page'
+
+export default function KakaoCallbackPage() {
+  return <OAuthCallbackPage providerOverride="kakao" />
+}